requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
pydantic==2.5.0
python-multipart==0.0.6

# HTTP requests for Azure Translator and Daily.co
requests==2.31.0

# Text-to-Speech (optional - can be added later)
gTTS==2.4.0

# Supabase for authentication
supabase==2.3.4

# JWT for token handling
pyjwt==2.8.0

# Utilities
python-dotenv==1.0.0